requirements.txt
streamlit==1.47.1
streamlit-lottie==0.0.5
requests==2.32.4
openai-whisper==20240930
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.1+cpu
torchaudio==2.3.1+cpu
numpy==2.2.6
groq==0.4.1
pydub==0.25.1
gTTS==2.5.4
python-dotenv==1.0.0
flask
gym
stable-baselines3
numpy
firebase-admin==6.2.0
flask-cors
requests
python-dotenv
